refactor: 입력 방법을 의존성 주입받게 됨에 따라 필요없는 인터페이스-구현 구조 제거
