--- src/main/java/lotto/input/UserInputPort.java
@@ -1,8 +1,0 @@
-package lotto.input;
-
-import lotto.domain.LottoAnswer;
-
-public interface UserInputPort {
-    int getLottoBuyPrice();
-    LottoAnswer getLottoAnswer();
-}
